+++ src/lib/pages/validator-details/components/validator-overview/index.tsx
@@ -4,11 +4,10 @@ import type Big from "big.js";
 import { VotingPowerChart } from "../bonded-token-changes/VotingPowerChart";
 import { Performance } from "../performance";
 import { RecentBlocksSection } from "../performance/RecentBlocksSection";
-import { ProposedBlocksTable } from "../tables/ProposedBlocksTable";
-import { VotedProposalsTable } from "../tables/VotedProposalsTable";
+import { ProposedBlocksTable } from "../tables/proposed-blocks";
+import { VotedProposalsTable } from "../tables/voted-proposals";
 import { useMobile } from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
-import { EmptyState } from "lib/components/state";
 import type { AssetInfos, Option, ValidatorAddr } from "lib/types";
 
 import { ValidatorDescription } from "./ValidatorDescription";
@@ -45,8 +44,21 @@ export const ValidatorOverview = ({
 }: ValidatorOverviewProps) => {
   const isMobile = useMobile();
 
-  return isActive && !isJailed ? (
+  return (
     <Flex direction="column" gap={{ base: 4, md: 6 }} pt={6}>
+      {(!isActive || isJailed) && (
+        <Alert variant="error" gap={2} my={6}>
+          <CustomIcon
+            name="alert-circle-solid"
+            boxSize={4}
+            color="error.main"
+          />
+          <AlertDescription>
+            This validator is currently {isJailed ? "jailed" : "inactive"}. The
+            information displayed reflects the latest available data.
+          </AlertDescription>
+        </Alert>
+      )}
       <ValidatorDescription details={details} />
       <Flex gap={{ base: 4, md: 6 }} direction={{ base: "column", md: "row" }}>
         <VotingPowerOverview
@@ -76,23 +88,10 @@ export const ValidatorOverview = ({
         validatorAddress={validatorAddress}
         onViewMore={onSelectPerformance}
       />
-      <VotedProposalsTable onViewMore={onSelectVotes} />
-    </Flex>
-  ) : (
-    <>
-      <Alert variant="error" gap={2} my={6}>
-        <CustomIcon name="alert-circle-solid" boxSize={4} color="error.main" />
-        <AlertDescription>
-          This validator is currently {isJailed ? "jailed" : "inactive"}. The
-          information displayed reflects the latest available data.
-        </AlertDescription>
-      </Alert>
-      <ValidatorDescription details={details} />
-      <EmptyState
-        message="This validator has recently begun their duties. Let's extend our best wishes for their success in their role. 😊"
-        imageVariant="empty"
-        withBorder
+      <VotedProposalsTable
+        validatorAddress={validatorAddress}
+        onViewMore={onSelectVotes}
       />
-    </>
+    </Flex>
   );
 };