+++ src/lib/types/validator.ts
@@ -5,6 +5,7 @@ import { formatUrl } from "lib/utils/formatter/url";
 
 import { zBechAddr20, zValidatorAddr } from "./addrs";
 import { zBig } from "./big";
+import { zRatio } from "./currency";
 import type { Ratio } from "./currency";
 
 export const zValidator = z
@@ -29,7 +30,7 @@ export const zValidatorData = z
     identity: z.string(),
     moniker: z.string(),
     details: z.string(),
-    commission_rate: z.coerce.number(),
+    commission_rate: zRatio(z.coerce.number()),
     is_jailed: z.boolean(),
     is_active: z.boolean(),
     voting_power: zBig,
@@ -58,3 +59,9 @@ export type ComputedUptime = {
   missedRatio: Ratio<number>;
   uptimeRatio: Ratio<number>;
 };
+
+export enum SlashingEvent {
+  Unjailed = "Unjailed",
+  Jailed = "Jailed",
+  Slashed = "Slashed",
+}